requirements.txt
certifi==2020.12.5
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
Flask==1.1.2
future==0.18.2
gym==0.18.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.5
Pillow==7.2.0
pyactup==1.0.9
pyglet==1.5.0
pylru==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
Werkzeug==1.0.1
